# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: M15-SchoolTech API — Plateforme de gestion scolaire
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rendez-vous/liste:
    get:
      operationId: listerRendezVous
      tags: [rendez-vous]
      summary: Lister les rendez-vous selon rôle
      security: [{ bearerAuth: [] }]
      parameters:
        - name: statut
          in: query
          schema: { type: string }
        - name: date_debut
          in: query
          schema: { type: string }
        - name: date_fin
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Liste des rendez-vous
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RendezVousListeResponse"
  /rendez-vous/demander:
    post:
      operationId: demanderRendezVous
      tags: [rendez-vous]
      summary: Demander un rendez-vous
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemanderRendezVousInput"
      responses:
        "201":
          description: Demande créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RendezVousItemResponse"
  /rendez-vous/{id}/confirmer:
    put:
      operationId: confirmerRendezVous
      tags: [rendez-vous]
      summary: Confirmer un rendez-vous
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmerRendezVousInput"
      responses:
        "200":
          description: Rendez-vous confirmé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RendezVousItemResponse"
  /rendez-vous/{id}/annuler:
    put:
      operationId: annulerRendezVous
      tags: [rendez-vous]
      summary: Annuler un rendez-vous
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Rendez-vous annulé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RendezVousItemResponse"
  /rendez-vous/{id}/terminer:
    put:
      operationId: terminerRendezVous
      tags: [rendez-vous]
      summary: Marquer un rendez-vous comme terminé
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Rendez-vous terminé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RendezVousItemResponse"

  # ── Annonces ──────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RendezVousItem:
      type: object
      properties:
        id: { type: string }
        etablissement_id: { type: string }
        parent_id: { type: string }
        parent_nom: { type: string }
        parent_prenoms: { type: string }
        professeur_id: { type: string, nullable: true }
        professeur_nom: { type: string, nullable: true }
        directeur_id: { type: string, nullable: true }
        eleve_id: { type: string }
        eleve_nom: { type: string }
        eleve_prenoms: { type: string }
        motif: { type: string }
        date_rdv: { type: string }
        heure_rdv: { type: string }
        duree_minutes: { type: integer }
        statut: { type: string, enum: [en_attente, confirme, annule, termine] }
        lieu: { type: string, nullable: true }
        notes_rdv: { type: string, nullable: true }
        created_at: { type: string }
    RendezVousItemResponse:
      type: object
      properties:
        rdv:
          $ref: "#/components/schemas/RendezVousItem"
    RendezVousListeResponse:
      type: object
      required: [rdvs]
      properties:
        rdvs:
          type: array
          items:
            $ref: "#/components/schemas/RendezVousItem"
    DemanderRendezVousInput:
      type: object
      required: [eleve_id, motif, date_rdv, heure_rdv]
      properties:
        professeur_id: { type: string }
        directeur_id: { type: string }
        eleve_id: { type: string }
        motif: { type: string }
        date_rdv: { type: string }
        heure_rdv: { type: string }
        duree_minutes: { type: integer }
    ConfirmerRendezVousInput:
      type: object
      properties:
        lieu: { type: string }
        notes_rdv: { type: string }

    # ── Annonces Schemas ────────────────────────────────────────
...